requirements.txt
django
djangorestframework
sshtunnel
mysql-connector-python
mysqlclient